[RISCV]Generate target attribute in attribute section of object file when assemble .s file
When there is not RISCV arch attribute in assembly .s file, it will not generate arch attribute in obj file. LLVM asm parser does not use subtarget feature bits to generate appropriate attributes. So instructions can not be recognized in disassemble process (llvm-objdump).

For example,

```
;t.s

fadd.s ft1,ft2,ft3
```

./bin/llvm-mc < t.s -mattr=+f -filetype=obj  |./bin/llvm-objdump -dr -

```
0000000000000000 <.text>:
       0: d3 70 31 00  	<unknown>
```

Emit target attribute at the constructor of RISCV asm parser to enable the functionality of attribute section even in assemble process.

It's a WIP patch to discuss and see whether the direction is right.

Differential Revision: https://reviews.llvm.org/D109587