[RISCV][RFC] Add Clang support for RISC-V overlay system
This is the Clang component of a change to add support for a proposed 'overlay' system for RISC-V to LLVM.

This adds two new attributes, __attribute__((overlaydata)), and __attribute__((overlaycall)), which are used to mark functions or global data as only accessible through the overlay engine. Internally they are converted to 'overlay-data' and 'overlay-call' attributes in LLVM IR.

This change also adds the option `-fcomrv` to enable use of the overlay system. This has the effect of reserving registers x28, x29, x30 and x31 for exclusive used by the overlay engine.

Differential Revision: https://reviews.llvm.org/D109372